fix dynamic

+++ wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/Dynamic/Flow_2_Noisy_Vehicles.groovy
@@ -154,9 +154,9 @@ def exec(Connection connection, input) {
     // Get every inputs
     // -------------------
 
-    double duration = 60
+    int duration = 60
     if (input['duration']) {
-        duration = Double.valueOf(input['duration'] as String)
+        duration = Integer.valueOf(input['duration'] as String)
     }
 
     int timestep = 1
